requirements.txt
fastapi==0.115.6
uvicorn==0.32.0
SQLAlchemy==2.0.36
python-dotenv==1.0.1
websockets==14.1
